refactor: refactor network functions into common library
- Move the `GetLocalIpAddresses` function to `common/io_util.cc`
- Add `common/io_util.h` header to `cluster/cluster.h`
- Use `util::GetLocalIpAddresses()` instead of the previously defined function in `cluster.cc`
